+++ test/libsolidity/syntaxTests/nameAndTypeResolution/526_fallback_marked_external_v050.sol
@@ -1,4 +1,0 @@
-pragma experimental "v0.5.0";
-contract C {
-    function () external { }
-}