+++ test/libsolidity/syntaxTests/nameAndTypeResolution/527_fallback_marked_internal_v050.sol
@@ -1,6 +1,0 @@
-pragma experimental "v0.5.0";
-contract C {
-    function () internal { }
-}
-// ----
-// TypeError: (47-71): Fallback function must be defined as "external".