+++ test/libsolidity/syntaxTests/nameAndTypeResolution/528_fallback_marked_private_v050.sol
@@ -1,6 +1,0 @@
-pragma experimental "v0.5.0";
-contract C {
-    function () private { }
-}
-// ----
-// TypeError: (47-70): Fallback function must be defined as "external".