+++ test/libsolidity/syntaxTests/nameAndTypeResolution/529_fallback_marked_public.sol
@@ -2,3 +2,4 @@ contract C {
     function () public { }
 }
 // ----
+// TypeError: (17-39): Fallback function must be defined as "external".